+++ packages/shared/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./formatting.js";
 export * from "./phase1.js";
 export * from "./plaud.js";
+export * from "./protocol.js";
 export * from "./runtime.js";